+++ packages/flame/lib/components.dart
@@ -18,7 +18,7 @@ export 'src/components/nine_tile_box_component.dart';
 export 'src/components/parallax_component.dart';
 export 'src/components/particle_component.dart';
 export 'src/components/position_component.dart';
-export 'src/components/positioning_type.dart';
+export 'src/components/position_type.dart';
 export 'src/components/shape_component.dart';
 export 'src/components/sprite_animation_component.dart';
 export 'src/components/sprite_animation_group_component.dart';
