+++ packages/flame/lib/src/components/position_type.dart
